split database.js into several files
