+++ src/backend/database/index.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const sequelize = require('./sequelize');
+const connectDatabase = require('./connect');
+const { User, Room, Message, RoomMember } = require('../models');
+
+module.exports = {
+  sequelize,
+  connectDatabase,
+  User,
+  Room,
+  Message,
+  RoomMember,
+};
