+++ src/backend/models/RoomMember.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const sequelize = require('../database/sequelize');
+
+const RoomMember = sequelize.define(
+  'RoomMember',
+  {},
+  {
+    tableName: 'room_members',
+  },
+);
+
+module.exports = RoomMember;
