+++ src/backend/models/index.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const { User, Room, Message, RoomMember } = require('./associations');
+
+module.exports = {
+  User,
+  Room,
+  Message,
+  RoomMember,
+};
